requirements/dev.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.6
psycopg==3.1.19
psycopg-binary==3.1.19
psycopg-pool==3.2.2
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
